extract base
