+++ src/TypealizR/GeneratedSourceFile.cs
@@ -3,7 +3,7 @@ using Microsoft.CodeAnalysis;
 
 namespace TypealizR;
 
-internal class GeneratedSourceFile
+public class GeneratedSourceFile
 {
 	public GeneratedSourceFile(string fileName, string content, IEnumerable<Diagnostic> diagnostics)
 	{
