+++ src/TypealizR/GeneratorOptions.cs
@@ -8,7 +8,7 @@ using TypealizR.Diagnostics;
 
 namespace TypealizR;
 
-internal sealed class GeneratorOptions
+public sealed class GeneratorOptions
 {
 	public const string MSBUILD_PROJECT_DIRECTORY = "build_property.msbuildprojectdirectory";
 	public const string PROJECT_DIR = "build_property.projectdir";
