+++ src/TypealizR/StringLocalizerSourceGenerator.cs
@@ -2,60 +2,15 @@
 using System.IO;
 using System.Linq;
 using Microsoft.CodeAnalysis;
-using TypealizR.Diagnostics;
 using TypealizR.Extensibility;
 using TypealizR.Builder;
 
 namespace TypealizR;
 
 [Generator]
-public partial class StringLocalizerSourceGenerator : IIncrementalGenerator
+public class StringLocalizerSourceGenerator : ResxFileSourceGeneratorBase
 {
-	public void Initialize(IncrementalGeneratorInitializationContext context)
-	{
-		var optionsProvider = context.AnalyzerConfigOptionsProvider.Select((x, cancel) => GeneratorOptions.From(x.GlobalOptions));
-		var resxFilesProvider = context.AdditionalTextsProvider.Where(static x => x.Path.EndsWith(".resx"));
-		var monitoredFiles = resxFilesProvider.Collect().Select((x, cancel) => RessourceFile.From(x));
-		
-		context.RegisterSourceOutput(monitoredFiles
-			.Combine(optionsProvider)
-			.Combine(context.CompilationProvider),
-			(ctxt, source) =>
-			{
-				var files = source.Left.Left;
-				var options = source.Left.Right;
-				var compilation = source.Right;
-
-				foreach (var file in files)
-				{
-					AddExtensionClassFor(ctxt, options, compilation, file);
-				}
-			});
-	}
-
-	private void AddExtensionClassFor(SourceProductionContext ctxt, GeneratorOptions options, Compilation compilation, RessourceFile file)
-	{
-		if (options.ProjectDirectory == null || !(options.ProjectDirectory.Exists))
-		{
-			ctxt.ReportDiagnostic(DiagnosticsFactory.TargetProjectRootDirectoryNotFound_0001());
-			return;
-		}
-
-		if (!file.Entries.Any())
-		{
-			return;
-		}
-
-		var generatedClass = GenerateExtensionClassFor(options.ProjectDirectory, options.RootNamespace, compilation, file, options.SeverityConfig);
-
-		ctxt.AddSource(generatedClass.FileName, generatedClass.Content);
-		foreach (var diagnostic in generatedClass.Diagnostics)
-		{
-			ctxt.ReportDiagnostic(diagnostic);
-		}
-	}
-
-	private GeneratedSourceFile GenerateExtensionClassFor(DirectoryInfo projectDir, string rootNamepsace, Compilation compilation, RessourceFile file, IDictionary<string, DiagnosticSeverity> severityConfig)
+	protected override GeneratedSourceFile GenerateSourceFileFor(DirectoryInfo projectDir, string rootNamepsace, Compilation compilation, RessourceFile file, IDictionary<string, DiagnosticSeverity> severityConfig)
 	{
 		var builder = new ClassBuilder(file.FullPath, severityConfig);
